Camera moving to the right place
