+++ Assets/Scripts/UI/3D UI Menu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cde74503705e55146bebeb027ab4d408
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
