+++ Assets/Scripts/UI/3D UI Menu/CustomizationMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 212fe104a35b88a48b595c47e1f012d3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
