+++ Assets/Scripts/UI/3D UI Menu/ExitMenu.cs
@@ -1,0 +1,24 @@
+using UnityEngine;
+
+public class ExitMenu : MonoBehaviour
+{
+    [Header("Textures")]
+    [SerializeField] private Texture selectedTexturel;
+    [SerializeField] private Texture idleTexture;
+
+    private void OnMouseDown()
+    {
+        MainMenu.Instance.Exit();
+        EventManager.TriggerEvent(GenericEvents.ButtonSound);
+    }
+
+    private void OnMouseEnter()
+    {
+        MainMenu.Instance.ChangeMenuMaterial(transform, selectedTexturel);
+    }
+
+    private void OnMouseExit()
+    {
+        MainMenu.Instance.ChangeMenuMaterial(transform, idleTexture);
+    }
+}
