+++ Assets/Scripts/UI/3D UI Menu/OptionsTab.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+
+public class OptionsTab : MonoBehaviour
+{
+    [Header("Tab")]
+    [SerializeField] private TabType tabType;
+
+    private void OnMouseDown()
+    {
+        MainMenu.Instance.SelectTab(tabType);
+        EventManager.TriggerEvent(GenericEvents.ButtonSound);
+    }
+}
