+++ Assets/Scripts/UI/3D UI Menu/PlayMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e1a0beeafef9eb94ca36d861b37980e5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
