--- Assets/Scripts/UI/CustomizationMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 212fe104a35b88a48b595c47e1f012d3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
