--- Assets/Scripts/UI/EditMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e4e790102a0d935488e75d7ce5107200
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
