--- Assets/Scripts/UI/ExitMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b4a6a8c9f281e57468e5bc895d9dee85
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
