--- Assets/Scripts/UI/OptionsMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1b28fe162a477a840aa013e1058d20db
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
