--- Assets/Scripts/UI/OptionsTab.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7bc876c9e15d4cd47bc3d3d572b95542
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
